разбивает main.js на модули
